requirements.txt
fastapi
uvicorn
python-multipart
python-dotenv
google-generativeai
pydantic
requests
pypdf
gunicorn
